Implemented run id=project id=curriculum folder name

+++ src/main/java/org/wise/portal/dao/project/ProjectDao.java
@@ -151,5 +151,5 @@ public interface ProjectDao<T extends Project> extends SimpleDao<T> {
 
   List<Project> getAllSharedProjects();
 
-  int getMaxProjectId();
+  long getMaxProjectId();
 }
